Move GlobalCache* to :persistentCache
