+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/state/SplitResourceSnapshotterCacheService.java
@@ -16,9 +16,9 @@
 
 package org.gradle.api.internal.changedetection.state;
 
+import org.gradle.cache.GlobalCacheLocations;
 import org.gradle.internal.hash.HashCode;
 import org.gradle.internal.snapshot.RegularFileSnapshot;
-import org.gradle.internal.vfs.GlobalCacheLocations;
 
 /**
  * A {@link ResourceSnapshotterCacheService} that delegates to the global service for immutable files
