+++ subprojects/core/src/main/java/org/gradle/internal/classpath/CopyingClasspathFileTransformer.java
@@ -16,9 +16,9 @@
 
 package org.gradle.internal.classpath;
 
+import org.gradle.cache.GlobalCacheLocations;
 import org.gradle.internal.file.FileType;
 import org.gradle.internal.snapshot.CompleteFileSystemLocationSnapshot;
-import org.gradle.internal.vfs.GlobalCacheLocations;
 import org.gradle.util.GFileUtils;
 
 import java.io.File;
