+++ subprojects/core/src/test/groovy/org/gradle/internal/service/scopes/VirtualFileSystemServicesTest.groovy
@@ -19,11 +19,11 @@ package org.gradle.internal.service.scopes
 import org.gradle.api.internal.GradleInternal
 import org.gradle.api.internal.StartParameterInternal
 import org.gradle.api.internal.cache.StringInterner
+import org.gradle.cache.GlobalCacheLocations
 import org.gradle.initialization.RootBuildLifecycleListener
 import org.gradle.internal.event.ListenerManager
 import org.gradle.internal.hash.FileHasher
 import org.gradle.internal.nativeintegration.filesystem.FileSystem
-import org.gradle.internal.vfs.GlobalCacheLocations
 import org.gradle.internal.vfs.RoutingVirtualFileSystem
 import org.gradle.internal.vfs.VirtualFileSystem
 import org.gradle.internal.watch.vfs.WatchingAwareVirtualFileSystem
