+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/ArtifactCachesProvider.java
@@ -15,7 +15,7 @@
  */
 package org.gradle.api.internal.artifacts.ivyservice;
 
-import org.gradle.internal.vfs.GlobalCache;
+import org.gradle.cache.GlobalCache;
 
 import java.io.Closeable;
 import java.util.Optional;
