+++ subprojects/file-watching/file-watching.gradle.kts
@@ -31,6 +31,7 @@ dependencies {
 
     testImplementation(project(":processServices"))
     testImplementation(project(":resources"))
+    testImplementation(project(":persistentCache"))
     testImplementation(project(":buildOption"))
     testImplementation(testFixtures(project(":core")))
     testImplementation(testFixtures(project(":fileCollections")))
