+++ subprojects/plugins/src/main/java/org/gradle/api/internal/tasks/CompileServices.java
@@ -23,11 +23,11 @@ import org.gradle.api.internal.tasks.compile.incremental.cache.DefaultUserHomeSc
 import org.gradle.api.internal.tasks.compile.incremental.cache.UserHomeScopedCompileCaches;
 import org.gradle.api.invocation.Gradle;
 import org.gradle.cache.CacheRepository;
+import org.gradle.cache.GlobalCacheLocations;
 import org.gradle.cache.internal.InMemoryCacheDecoratorFactory;
 import org.gradle.initialization.JdkToolsInitializer;
 import org.gradle.internal.service.ServiceRegistration;
 import org.gradle.internal.service.scopes.AbstractPluginServiceRegistry;
-import org.gradle.internal.vfs.GlobalCacheLocations;
 import org.gradle.internal.vfs.VirtualFileSystem;
 
 public class CompileServices extends AbstractPluginServiceRegistry {
